requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
torch==2.0.0
torchvision==0.15.1
pillow==9.1.1
python-multipart
